# app/static/swagger.yaml
swagger: "2.0"
info:
  title: "Mechanic Shop API"
  description: "A comprehensive REST API for managing a mechanic shop with customers, vehicles, mechanics, service tickets, and inventory. Includes JWT authentication, rate limiting, and caching."
  version: "2.0.0"
  contact:
    name: "API Support"
    url: "https://github.com/KimiCubb/Mechanics-Shop-db"
host: "mechanics-shop-db.onrender.com"
basePath: ""
schemes:
  - "https"
consumes:
  - "application/json"
produces:
  - "application/json"
securityDefinitions:
  Bearer:
    type: "apiKey"
    name: "Authorization"
    in: "header"
    description: "JWT Bearer token. Format: Bearer <token>"
paths:
...
  /vehicles:
    post:
      tags:
        - "Vehicles"
      summary: "Create a new vehicle"
      description: "Register a new vehicle for a customer"
      parameters:
        - in: "body"
          name: "body"
          required: true
          schema:
            $ref: "#/definitions/VehicleCreate"
      responses:
        201:
          description: "Vehicle created successfully"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
              vehicle:
                $ref: "#/definitions/VehicleResponse"
        400:
          description: "Invalid input"
        404:
          description: "Customer not found"

    get:
      tags:
        - "Vehicles"
      summary: "Get all vehicles"
      description: "Retrieve all vehicles in the system"
      responses:
        200:
          description: "Vehicles retrieved successfully"
          schema:
            $ref: "#/definitions/VehiclesListResponse"
  /vehicles/{vehicle_id}:
    get:
      tags:
        - "Vehicles"
      summary: "Get a specific vehicle"
      description: "Retrieve details for a specific vehicle"
      parameters:
        - in: "path"
          name: "vehicle_id"
          type: "integer"
          required: true
      responses:
        200:
          description: "Vehicle retrieved successfully"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
              vehicle:
                $ref: "#/definitions/VehicleResponse"
        404:
          description: "Vehicle not found"

    put:
      tags:
        - "Vehicles"
      summary: "Update a vehicle"
      description: "Update vehicle information"
      parameters:
        - in: "path"
          name: "vehicle_id"
          type: "integer"
          required: true
        - in: "body"
          name: "body"
          schema:
            $ref: "#/definitions/VehicleCreate"
      responses:
        200:
          description: "Vehicle updated successfully"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
              vehicle:
                $ref: "#/definitions/VehicleResponse"
        400:
          description: "Invalid input"
        404:
          description: "Vehicle not found"

    delete:
      tags:
        - "Vehicles"
      summary: "Delete a vehicle"
      description: "Remove a vehicle from the system"
      parameters:
        - in: "path"
          name: "vehicle_id"
          type: "integer"
          required: true
      responses:
        200:
          description: "Vehicle deleted successfully"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
        404:
          description: "Vehicle not found"
  /vehicles/customer/{customer_id}:
    get:
      tags:
        - "Vehicles"
      summary: "Get vehicles for a customer"
      description: "Retrieve all vehicles belonging to a specific customer"
      parameters:
        - in: "path"
          name: "customer_id"
          type: "integer"
          required: true
      responses:
        200:
          description: "Vehicles retrieved successfully"
          schema:
            $ref: "#/definitions/CustomerVehiclesResponse"
        404:
          description: "Customer not found"
definitions:
...
  VehicleResponse:
    type: "object"
    properties:
      vehicle_id:
        type: "integer"
      customer_id:
        type: "integer"
      year:
        type: "integer"
      make:
        type: "string"
      model:
        type: "string"
      vin:
        type: "string"
      license_plate:
        type: "string"
  VehicleCreate:
    type: "object"
    required:
      - customer_id
      - year
      - make
      - model
      - vin
    properties:
      customer_id:
        type: "integer"
        example: 1
      year:
        type: "integer"
        example: 2021
      make:
        type: "string"
        maxLength: 50
        example: "Honda"
      model:
        type: "string"
        maxLength: 50
        example: "Civic"
      vin:
        type: "string"
        maxLength: 17
        example: "12345ABCDE67890FG"
      license_plate:
        type: "string"
        maxLength: 20
        example: "ABC123"
  VehiclesListResponse:
    type: "object"
    properties:
      message:
        type: "string"
      count:
        type: "integer"
      vehicles:
        type: "array"
        items:
          $ref: "#/definitions/VehicleResponse"
  CustomerVehiclesResponse:
    type: "object"
    properties:
      message:
        type: "string"
      customer_id:
        type: "integer"
      count:
        type: "integer"
      vehicles:
        type: "array"
        items:
          $ref: "#/definitions/VehicleResponse"
...